Add hostname and extensionId to site metadata
The hostname is now used as a fallback in case the siteTitle is missing
from the metadata (as it would be for connections with extensions).

The artificial hostname set for internal connections has been renamed
from 'MetaMask' to 'metamask' because URL objects automatically
normalize hostnames to be all lower-case, and it was more convenient to
use a URL object so that the parameter would be the same type as used
for an untrusted connection.